+++ src/record/string.rs
@@ -3,7 +3,7 @@ use super::*;
 macro_rules! impl_encoder_for {
     [$($ty:ty),*] => {$(
         impl Encode for $ty {
-            #[inline] fn encode<const CONFIG: u8>(&self, c: &mut (impl Write + ?Sized)) -> io::Result<()> {
+            #[inline] fn encode<const CONFIG: u16>(&self, c: &mut (impl Write + ?Sized)) -> io::Result<()> {
                 encode_len!(self, c);
                 c.write_all(self.as_ref())
             }
@@ -21,7 +21,7 @@ macro_rules! read_slice {
 
 impl<'de> Decode<'de> for String {
     #[inline]
-    fn decode<const CONFIG: u8>(c: &mut &'de [u8]) -> Result<Self> {
+    fn decode<const CONFIG: u16>(c: &mut &'de [u8]) -> Result<Self> {
         let data = read_slice!(c)?;
         String::from_utf8(data.to_vec()).map_err(Error::from)
     }
@@ -29,7 +29,7 @@ impl<'de> Decode<'de> for String {
 
 impl<'de: 'a, 'a> Decode<'de> for &'a str {
     #[inline]
-    fn decode<const CONFIG: u8>(c: &mut &'de [u8]) -> Result<Self> {
+    fn decode<const CONFIG: u16>(c: &mut &'de [u8]) -> Result<Self> {
         let data = read_slice!(c)?;
         std::str::from_utf8(data).map_err(Error::from)
     }
@@ -37,7 +37,7 @@ impl<'de: 'a, 'a> Decode<'de> for &'a str {
 
 impl<'de: 'a, 'a> Decode<'de> for &'a [u8] {
     #[inline]
-    fn decode<const CONFIG: u8>(c: &mut &'de [u8]) -> Result<Self> {
+    fn decode<const CONFIG: u16>(c: &mut &'de [u8]) -> Result<Self> {
         read_slice!(c)
     }
 }
