+++ src/types/enumerate.rs
@@ -5,7 +5,7 @@ where
     T: Encode,
 {
     #[inline]
-    fn encode<const CONFIG: u8>(&self, c: &mut (impl Write + ?Sized)) -> io::Result<()> {
+    fn encode<const CONFIG: u16>(&self, c: &mut (impl Write + ?Sized)) -> io::Result<()> {
         match self {
             Some(val) => {
                 c.write_all(&[1])?;
@@ -18,7 +18,7 @@ where
 
 impl<'de, T: Decode<'de>> Decode<'de> for Option<T> {
     #[inline]
-    fn decode<const CONFIG: u8>(r: &mut &'de [u8]) -> Result<Self> {
+    fn decode<const CONFIG: u16>(r: &mut &'de [u8]) -> Result<Self> {
         Ok(match bool::decode::<CONFIG>(r)? {
             true => Some(T::decode::<CONFIG>(r)?),
             false => None,
@@ -32,7 +32,7 @@ where
     E: Encode,
 {
     #[inline]
-    fn encode<const CONFIG: u8>(&self, c: &mut (impl Write + ?Sized)) -> io::Result<()> {
+    fn encode<const CONFIG: u16>(&self, c: &mut (impl Write + ?Sized)) -> io::Result<()> {
         match self {
             Ok(val) => {
                 c.write_all(&[1])?;
@@ -52,7 +52,7 @@ where
     E: Decode<'de>,
 {
     #[inline]
-    fn decode<const CONFIG: u8>(c: &mut &'de [u8]) -> Result<Self> {
+    fn decode<const CONFIG: u16>(c: &mut &'de [u8]) -> Result<Self> {
         Ok(match bool::decode::<CONFIG>(c)? {
             true => Ok(T::decode::<CONFIG>(c)?),
             false => Err(E::decode::<CONFIG>(c)?),
