+++ src/types/other.rs
@@ -3,7 +3,7 @@ use std::ops::{Range, RangeInclusive};
 
 impl<T: Encode> Encode for Range<T> {
     #[inline]
-    fn encode<const CONFIG: u8>(&self, c: &mut (impl Write + ?Sized)) -> io::Result<()> {
+    fn encode<const CONFIG: u16>(&self, c: &mut (impl Write + ?Sized)) -> io::Result<()> {
         self.start.encode::<CONFIG>(c)?;
         self.end.encode::<CONFIG>(c)
     }
@@ -11,7 +11,7 @@ impl<T: Encode> Encode for Range<T> {
 
 impl<'de, T: Decode<'de>> Decode<'de> for Range<T> {
     #[inline]
-    fn decode<const CONFIG: u8>(c: &mut &'de [u8]) -> Result<Self> {
+    fn decode<const CONFIG: u16>(c: &mut &'de [u8]) -> Result<Self> {
         let start = T::decode::<CONFIG>(c)?;
         let end = T::decode::<CONFIG>(c)?;
         Ok(start..end)
@@ -20,7 +20,7 @@ impl<'de, T: Decode<'de>> Decode<'de> for Range<T> {
 
 impl<T: Encode> Encode for RangeInclusive<T> {
     #[inline]
-    fn encode<const CONFIG: u8>(&self, c: &mut (impl Write + ?Sized)) -> io::Result<()> {
+    fn encode<const CONFIG: u16>(&self, c: &mut (impl Write + ?Sized)) -> io::Result<()> {
         self.start().encode::<CONFIG>(c)?;
         self.end().encode::<CONFIG>(c)
     }
@@ -28,7 +28,7 @@ impl<T: Encode> Encode for RangeInclusive<T> {
 
 impl<'de, T: Decode<'de>> Decode<'de> for RangeInclusive<T> {
     #[inline]
-    fn decode<const CONFIG: u8>(c: &mut &'de [u8]) -> Result<Self> {
+    fn decode<const CONFIG: u16>(c: &mut &'de [u8]) -> Result<Self> {
         let start = T::decode::<CONFIG>(c)?;
         let end = T::decode::<CONFIG>(c)?;
         Ok(start..=end)
